refactor: move task modal out of todo list
