+++ app/app.vue
@@ -62,8 +62,10 @@
         <NuxtPage />
       </main>
     </div>
-  </div>
-</template>
+      <CategoryModal />
+      <TaskModal />
+    </div>
+  </template>
 
 <script setup lang="ts">
 import { useRouter, useRoute } from 'vue-router'
@@ -73,7 +75,9 @@ import DatePicker from 'vue-datepicker-next'
 import 'vue-datepicker-next/index.css'
 import { format } from 'date-fns'
 import CategoryList from './components/CategoryList.vue'
-import { textColor } from './utils/color'
+  import CategoryModal from './components/CategoryModal.vue'
+  import TaskModal from './components/TaskModal.vue'
+  import { textColor } from './utils/color'
 
 interface Todo {
   date: string
